+++ src/PingFlud.Application/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("PingFlud.Application.Tests")]
+[assembly: InternalsVisibleTo("PingFlud")]